Extract crawler logic and filter logic to new class file
