+++ src/main/kotlin/net/purefunc/generate/Main.kt
@@ -1,71 +1,14 @@
 package net.purefunc.generate
 
-import java.io.BufferedReader
+import net.purefunc.generate.util.EmojiReader
+import net.purefunc.generate.util.ValidEmojiCollector
 import java.io.File
 import java.io.FileOutputStream
-import java.io.InputStreamReader
-import java.net.URL
 import java.util.Locale
 
 fun main() {
-    // collect emoji lines
-    // 1F636 200D 1F32B FE0F                                  ; fully-qualified     # 😶‍🌫️ E13.1 face in clouds
-    var flag = false
-    val lines = mutableListOf<String>()
-    val url = URL("https://unicode.org/Public/emoji/15.0/emoji-test.txt")
-    val reader = BufferedReader(InputStreamReader(url.openConnection().getInputStream()))
-    reader.useLines { readLines ->
-        readLines.forEach { line ->
-            if (line.isEmpty()) flag = false
-            if (flag) lines.add(line)
-            if (line.startsWith("# subgroup: ")) flag = true
-        }
-    }
-
-    val bigEnum = lines.filter {
-        it.contains("fully-qualified")
-    }.map {
-        it.split(" ")
-    }.map { elements ->
-        // [1F636, 200D, 1F32B, FE0F, , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , , ;, fully-qualified, , , , , #, 😶‍🌫️, E13.1, face, in, cloud]
-
-        val items = mutableListOf<String>()
-
-        val codePartIdxs = elements.mapIndexed { idx, element -> if (element == "" || element == ";") idx else -1 }
-        val codeIdx = codePartIdxs.first { idx -> idx != -1 }
-
-        (0 until codeIdx).forEach { i -> items.add(elements[i]) }
-        items.add(";")
-
-        val namePartIdxs =
-            elements.mapIndexed { idx, element -> if (element.startsWith("E") && element.contains(".")) idx else -1 }
-        val nameIdx = namePartIdxs.first { idx -> idx != -1 }
-
-        items.addAll(
-            // replace item name contains invalid char
-            elements.subList(nameIdx + 1, elements.size).map { str ->
-                str.replace("“", "")
-                    .replace("”", "")
-                    .replace("’", "")
-                    .replace("-", "_")
-                    .replace(":", "")
-                    .replace(".", "")
-                    .replace("!", "")
-                    .replace("(", "")
-                    .replace(")", "")
-                    .replace("1st", "first")
-                    .replace("2nd", "second")
-                    .replace("3rd", "third")
-                    .replace("package", "packages")
-                    .replace("#", "hash")
-                    .replace("*", "star")
-                    .replace(",", "comma")
-                    .replace("&", "and")
-            }
-        )
-
-        items
-    }
+    val roughList = EmojiReader("https://unicode.org/Public/emoji/15.0/emoji-test.txt").readTargetUrl()
+    val bigEnum = ValidEmojiCollector(roughList).filter()
 
     // all enum in one .kt is will exceed jvm limit 64K
     val pageCount = 1000
